refactor(debug_server): extract metrics + diagnostics + device control to debug_server_metrics (refs #332)
Wave 23b · 18th (and final) per-family extract.  All 11 small read-only /
control endpoints that didn't fit any other family bucket move from
debug_server.c into a new sibling module debug_server_metrics.{c,h},
plus the /heap/probe-csv pool-probe relay (handler stays in pool_probe.c
but its uri_t + register call live with the metrics family for cohesion).

Endpoints moved (all bearer-auth):
  GET  /tasks            — FreeRTOS task snapshot
  GET  /logs/tail?n=     — last N lines from obs log ring
  GET  /battery          — voltage/current/power/percent/charging
  POST /display/brightness?pct=
  GET  /audio            — current volume + mic_mute
  POST /audio?action=    — volume|mute control
  GET  /metrics          — Prometheus text format
  GET  /events?since=ms  — obs events ring as JSON
  GET  /heap/history?n=  — heap sampling history JSON
  GET  /heap/probe-csv   — pool-probe CSV relay
  GET  /keyboard/layout  — live LVGL keyboard key positions
  POST /tool_log/push    — synthesise tool_log event for harness
  GET  /net/ping?host=&port= — non-blocking TCP probe (2s timeout)

Helpers moved alongside:
  - url_pct_decode_inplace (last consumer in debug_server.c was
    tool_log_push_handler — debug_server.c no longer needs its own copy)

debug_server.c is now down to ~700 LOC owning ONLY the core surface:
httpd lifecycle + bearer auth init + send_json_resp helper +
/info /index /selftest core handlers + the 18 family register calls.

E2E verification on physical Tab5 192.168.1.90:
  - story_smoke 14/14 in 67s
  - GET /tasks /battery /audio /metrics /events /heap/history
        /keyboard/layout all return 200 with payloads
  - GET /net/ping?host=192.168.1.91&port=3502 → ok=true, elapsed_ms=2

debug_server.c LOC: 1,272 → ~700 (-572).
debug_server_metrics.c: 0 → 540.
